Move the Auspice client's Charon endpoints into a separate directory
A reorganization of files that aims to establish a clearer structure for
adding more endpoints for other nextstrain.org APIs.
