+++ src/endpoints/charon/getAvailable.js
@@ -1,7 +1,7 @@
-const utils = require("./utils");
+const utils = require("../../utils");
 const queryString = require("query-string");
-const {splitPrefixIntoParts, joinPartsIntoPrefix} = require("./utils/prefix");
-const metaSources = require("./metaSources");
+const {splitPrefixIntoParts, joinPartsIntoPrefix} = require("../../utils/prefix");
+const metaSources = require("../../metaSources");
 
 /* handler for /charon/getAvailable requests */
 const getAvailable = async (req, res) => {
