+++ src/endpoints/charon/getNarrative.js
@@ -1,8 +1,8 @@
 const fetch = require('node-fetch');
 const {BadRequest, NotFound} = require("http-errors");
 
-const utils = require("./utils");
-const {splitPrefixIntoParts} = require("./utils/prefix");
+const utils = require("../../utils");
+const {splitPrefixIntoParts} = require("../../utils/prefix");
 
 const getNarrative = async (req, res) => {
   const query = req.query;
